% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\name{nations}
\alias{nations}
\title{nations data}
\usage{
data(nations)
}


\description{
The array "R" is a 14 × 14 × 56 binary tensor consisting of 56 political relations of 14 countries between 1950 and 1965. The tensor entry indicates the presence or absence of a political action, such as “treaties”, “sends tourists to”, between the nations. Please set the diagonal elements Y(i,i,k) = 0 in the analysis.

The matrix "cov" is a 14 × 6 matrix describing a few important country attributes, e.g. whether a nation is actively involved in medicine NGO, law NGO, or belongs to a catholic nation, etc.
}

\references{
Nickel M, Tresp V, Kriegel H P. A three-way model for collective learning on multi-relational data[C]//Icml. 2011, 11: 809-816. URL: https://arxiv.org/pdf/1306.2084.pdf
}
\format{ One 14-14-56 binary array named "R"; one 14-6 matrix named "cov".}

\keyword{datasets}
