% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sele_rank.R
\name{sele_rank}
\alias{sele_rank}
\title{Rank selection}
\usage{
sele_rank(tsr, X_covar1 = NULL, X_covar2 = NULL, X_covar3 = NULL,
  rank_range, Nsim = 10, cons = "non", dist)
}
\arguments{
\item{tsr}{response tensor with 3 modes}

\item{X_covar1}{covariate on first mode}

\item{X_covar2}{covariate on second mode}

\item{X_covar3}{covariate on third mode}

\item{rank_range}{a matrix containing rank candidates on each row}

\item{Nsim}{max number of iterations if update does not convergence}

\item{cons}{the constraint method, "non" for without constraint, "vanilla" for global scale down at each iteration,

                   "penalty" for adding log-barrier penalty to object function.}

\item{dist}{distribution of response tensor, see "details"}
}
\value{
a list containing the following:

                   \code{rank} selected rank with minimal BIC

                   \code{result}  a matrix containing rank candidate and its loglikelihood and BIC on each row
}
\description{
Estimate the Tucker rank of coefficient tensor based BIC criterion. The choice of BIC
 aims to balance between the goodness-of-fit for the data and the degree of freedom in the population model.
}
\details{
For rank selection, recommend using non-constraint version.
            
            \code{dist} specifies three distributions of response tensor: binary, poisson and normal distributions.
}
